Add boards.yaml template for CIP LAVA slave labs
This is a bare minimum template for the creation of a CIP LAVA slave
lab. To get a lab working using this template, all the <variables> need
to be changed to actual values and if a user wishes to do so, they can
add more DUTs to the "boards" section.

Signed-off-by: Patryk Mungai <patryk.mungai-ndungu.kx@renesas.com>