Update RZ/G1[EM] healthchecks to use cip-latest
CIP's Amazon S3 bucket directory structure has changed. This commit
updates the urls for the kernels, dtbs and rfs'.

The commit also includes a notification section to send emails to the
mailing list at the end of each healthcheck. It also removes the
redundant .empty file in the healthcheck repository.

Signed-off-by: Patryk Mungai <patryk.mungai-ndungu.kx@renesas.com>